🏗 folder :  불필요한 레이아웃 삭제
