+++ src/app/(Main)/feed/page.tsx
@@ -2,8 +2,8 @@
 
 import { QueryAsyncBoundary } from '@suspensive/react-query';
 
-import FilterSheet from './components/FilterSheet';
 import FeedContent from './FeedContent';
+import FilterSheet from './FilterSheet';
 import { RejectedFallback } from '@/components/ErrorBoundary';
 import { Loading } from '@/components/Loading';
 import { PageAnimation } from '@/components/PageAnimation';
