--- src/app/(Sub)/layout.tsx
@@ -1,5 +1,0 @@
-import { StrictPropsWithChildren } from '@/types';
-
-export default function SubLayout({ children }: StrictPropsWithChildren) {
-  return <>{children}</>;
-}
